Path tracing working for small input in CUDA renderer.
git-svn-id: svn+ssh://jhoberock-dt/home/jhoberock/svnbase/jared/src/gotham@1512 3bb10773-34f4-0310-b88c-ec9d511ba0d9

+++ shading/ScatteringDistributionFunction.h
@@ -31,7 +31,7 @@ class ScatteringDistributionFunction
     /*! \typedef ComponentIndex
      *  \brief Shorthand.
      */
-    typedef size_t ComponentIndex;
+    typedef unsigned int ComponentIndex;
 
     /*! Null destructor does nothing.
      */
